Implement AUTOFILL menu
Bug:

+++ content/public/android/java/src/org/chromium/content/browser/SelectionPopupController.java
@@ -92,6 +92,7 @@ public class SelectionPopupController extends ActionModeCallbackHelper {
     private final WebContents mWebContents;
     private final RenderCoordinates mRenderCoordinates;
     private ActionMode.Callback mCallback;
+    private ActionMode.Callback mNonSelectionCallback;
 
     // Selection rectangle in DIP.
     private final Rect mSelectionRect = new Rect();
@@ -205,6 +206,10 @@ public class SelectionPopupController extends ActionModeCallbackHelper {
         mCallback = callback;
     }
 
+    void setNonSelectionCallback(ActionMode.Callback callback) {
+        mNonSelectionCallback = callback;
+    }
+
     @Override
     public boolean isActionModeValid() {
         return mActionMode != null;
@@ -296,7 +301,7 @@ public class SelectionPopupController extends ActionModeCallbackHelper {
             return;
         }
 
-        if (!supportsFloatingActionMode() && !canPaste()) return;
+        if (!supportsFloatingActionMode() && (!canPaste() || mNonSelectionCallback == null)) return;
         destroyPastePopup();
         PastePopupMenuDelegate delegate = new PastePopupMenuDelegate() {
             @Override
@@ -334,7 +339,9 @@ public class SelectionPopupController extends ActionModeCallbackHelper {
         Context windowContext = mWindowAndroid.getContext().get();
         if (windowContext == null) return;
         if (supportsFloatingActionMode()) {
-            mPastePopupMenu = new FloatingPastePopupMenu(windowContext, mView, delegate);
+            // NonSelectionCallback has only been needed since Android O.
+            mPastePopupMenu = new FloatingPastePopupMenu(
+                    windowContext, mView, delegate, mNonSelectionCallback);
         } else {
             mPastePopupMenu = new LegacyPastePopupMenu(windowContext, mView, delegate);
         }
